Reform structure
Since the program is starting to work, I would like others to be able
to use it. I'm restructuring the code, so it's easier to install, and
it's easier for other people to get involved. I hope this is the last
major change that makes it harder to search the git tree.